Sign Up & Signin Done

+++ api/src/Application/Actions/Review/DeleteReviewAction.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace App\Application\Actions\Review;
 
+use Exception;
 use Psr\Http\Message\ResponseInterface as Response;
 
 class DeleteReviewAction extends ReviewAction
@@ -15,6 +16,11 @@ class DeleteReviewAction extends ReviewAction
         $id = (int) $this->resolveArg('id');
 
         $review = $this->reviewRepository->findReviewOfId($id);
+
+        if (!$this->isCustomer() ||
+            !$review->customer->is($this->getAuthCustomer())) {
+            throw new Exception('Unauthorized');
+        }
 
         $this->reviewRepository->destroy($id);
